feat : ViewModel 로직 수정

+++ GraphControl/Model/Graph.swift
@@ -10,5 +10,5 @@ import Foundation
 struct Graph {
     static let minValue = 0
     static let maxValue = 100
-    var value: Double
+    var value: Float
 }
